fixed linetracker

+++ project/linetracking_system/linetracker.py
@@ -9,7 +9,7 @@ class LineTracker(RobotMovement):
         self.color_sensor = color_sensor
         self.isLeft = False
 
-    def follow_line(self, base_power: int = 30, correction_factor: float = 10.0):
+    def follow_line(self, base_power: int = 30, correction_factor: int = 10):
         """
         Adjusts motor speeds based on sensor reading to follow a line.
